refactor: remove Parser module to reduce namespace
